bug #3319036: fix one of the (severe) SMS bugs where the wrong conversation is opened.
- the code in ConversationList.onListItemClicked() was extracting the conversation data
from the ConversationListItem view passed in. However, ListItem views can be recycled
when rebound, so the view we got could now be reassigned to a different list position
and contain different data.

Intead of depend on ConversationListItem to store the conversation data, always load
the cursor at the position clicked, and get the conversation data from the cursor.

Change-Id: Ic66d072c6e96de0082051500ea118545886df9db